Port to 64-bit linux.
* Makefile: Attempt to compensate for broken make
on linux 64.
* Makefile.unix: New file.
* lib/xdr.c: Silence warnings. Remove unnecessary
parentheses around TRUE/FALSE constants.
(xdr_int): Adapt when long bigger than 32 bits.
(xdr_u_int): Ditto.
(xdr_long): Ditto. NOTE: now returns FALSE when input
quantity is larger than can be represented using 4 bytes.
Use xdr_int64_t or xdr_hyper instead.
(xdr_u_long): Ditto. Use xdr_uint64_t or xdr_u_hyper.
(xdr_short, xdr_u_short): Formatting changes.
(xdr_char, xdr_u_char): Formatting changes.
(xdr_bool): Formatting changes.
(xdr_int32_t, xdr_u_int32_t, xdr_uint32_t): Use
XDR_[PUT|GET]INT32 instead of XDR_[PUT|GET]LONG.
(xdr_int16_t, xdr_u_int16_t, xdr_uint16_t): Use
XDR_[PUT|GET]INT32 instead of XDR_[PUT|GET]LONG.
(xdr_int8_t, xdr_u_int8_t, xdr_uint8_t): Use
XDR_[PUT|GET]INT32 instead of XDR_[PUT|GET]LONG.
(xdr_enum): Adapt when enum larger than 32 bits.
(xdr_opaque): Silence compiler warnings. Remove
unnecessary type-punning. Formatting changes.
(xdr_bytes, xdr_union): Formatting changes.
(xdr_string): Formatting changes. Fast return on
errors.
(xdr_int64_t, xdr_uint64_t, xdr_u_int64_t): Use int32_t
for temporary storage, not longs. Avoid unnecessary
array indexing. Use XDR_[PUT|GET]INT32 rather than
XDR_[PUT|GET]LONG.
* lib/xdr_mem.c: Silence warnings. Remove unnecessary
parentheses around TRUE/FALSE constants.
(xdrmem_[get|put]int32_[un]aligned): Four new functions.
(xdrmem_ops_[un]aligned): Add new functions to
associated structs.
(xdrmem_setpos): Simplify code for dealing with 64 bits.
* lib/xdr_private.h: include <endian.h>, not
<machine/endian.h> when appropriate.
* lib/xdr_reference.c: Silence warnings. Remove unnecessary
parentheses around TRUE/FALSE constants.
* lib/xdr_sizeof.c: Remove unnecessary parentheses around
TRUE/FALSE constants.  Other formatting changes.
(x_putint32): New function.
(xdr_sizeof): Add new typedef for x_[put|get]int32 members
of xdr_opts struct. Initialize those members.
* lib/xdr_stdio.c: Silence warnings. Remove unnecessary
parentheses around TRUE/FALSE constants.
(xdrstdio_[put|get]int32): New functions.
(xdrstdio_ops): Initialize x_[put|get]int32 members.
(xdrstdio_create): Cast to correct types.
(xdrstdio_getlong): Cast to correct types.
(xdrstdio_putlong): Use correct type for intermediate storage.
* rpc/xdr.h: Use faster implementation of RNDUP macro.
Rearrange function declarations. Add missing xdr_sizeof decl.
(struct xdr_ops): Add new members x_getint32 and x_putint32.
Remove unused member x_control.
(XDR_[PUT|GET]INT32): redefine to use xdr_ops members, rather
than static inline methods...
(xdr_[put|get]int32): Removed.
(XDR_DESTROY, xdr_destroy): Redefine to use canonical
do ... while (0) construction for ;-swallowing.
(XDR_CONTORL, xdr_control): Removed.
(xdr_rpc*): Removed unused rpc macros.
(IXDR_*): Rearrange inline macros, and add warning
against their use (they use xdr_long, which is not 64bit clean).
* src/test/test_common.h: Add missing limits.h.
* src/test/test_data.h (struct gnumbers): Use int32_t, not long.
* src/test/test_data.c: Ripple from struct gnumbers change.
* src/test/test_xdrs.c: Remove trailing spaces.
(test_core_xdr_array): Silence typepunning warnings by casting
to void.
* src/test/xdrmem_test.c (main): Don't test xdr_long and
xdr_u_long if 64bits. We know it is broken. Silence warnings.
(test_xdrmem_list): Use correct non-recursive xdr function.
* src/test/xdrstdio_test.c (main): Don't test xdr_long and
xdr_u_long if 64bits. We know it is broken. Silence warnings.

+++ src/test/test_data.h
@@ -36,8 +36,8 @@ extern const struct xdr_discrim test_union_dscrim[6];
 
 /* struct definitions and helper functions for xdr_reference */
 typedef struct gnumbers_ {
-  long g_assets;
-  long g_liabilities;
+  int32_t g_assets;
+  int32_t g_liabilities;
 } gnumbers_t;
 extern bool_t xdr_gnumbers_t (XDR *, gnumbers_t *);
 typedef struct pgn_ {
